Make clientID unique
Mosquitto would close all "old" connections for the same clientID on connect.
Which resulted in 2 sockets continously connecting because they would be kicked by the other socket again.

In order to be able to identify them and in order to use the same image for all sockets, i used the MAC address
instead of something runtime-random or even compile-time-random