add support for TH10 & TH16 (humidity and temperature sensor)
If you bought a Sonoff TH10 or TH16 you can now use the temperature and humidity sensor. I testet it with AM2301 which basically is a DHT22. 

The 1-wire is connected to GPIO14. Both values are shown on the default web page. To use ist change #define IS_TH10_16 from 0 to 1.